refactor: main: Rename common >> app
