+++ ewm-main/src/main/java/ru/practicum/ewm/app/config/ObjectMapperConfig.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.common.config;
+package ru.practicum.ewm.app.config;
 
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
