+++ ewm-main/src/main/java/ru/practicum/ewm/app/exception/handler/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.common.exception.handler;
+package ru.practicum.ewm.app.exception.handler;
 
 import lombok.Value;
