+++ ewm-main/src/main/java/ru/practicum/ewm/category/model/CategoryMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.ewm.category.model;
 
 import org.mapstruct.Mapper;
-import ru.practicum.ewm.common.BaseMapper;
+import ru.practicum.ewm.app.BaseMapper;
 
 @Mapper(componentModel = "spring")
 public interface CategoryMapper extends BaseMapper<Category, CategoryDto> {
