+++ ewm-main/src/main/java/ru/practicum/ewm/user/model/UserMapper.java
@@ -1,7 +1,7 @@
 package ru.practicum.ewm.user.model;
 
 import org.mapstruct.Mapper;
-import ru.practicum.ewm.common.BaseMapper;
+import ru.practicum.ewm.app.BaseMapper;
 
 @Mapper(componentModel = "spring")
 public interface UserMapper extends BaseMapper<User, UserDto> {
